(bug) Set SveltosCluster.Spec.KubeconfigKeyName
When a cluster is registered with Sveltos, a Secret containing the cluster's
kubeconfig is created.
The kubeconfig is stored under the key __kubeconfig__ within this Secret.

If Sveltos renews the token, the key is updated to __re-kubeconfig__, and
the `SveltosCluster.Spec.KubeconfigKeyName` is updated to reflect this change.

If you run the `sveltosctl register cluster` command again, the Secret's data
is reset, and the cluster's kubeconfig is once again stored under the __kubeconfig__ key.

Previously, this would cause issues because the `SveltosCluster.Spec.KubeconfigKeyName`
was not updated back to __kubeconfig__.
This PR resolves that problem by ensuring that `SveltosCluster.Spec.KubeconfigKeyName`
is correctly updated to match the key in the Secret, preventing Sveltos components from
failing to find the cluster's kubeconfig.